refactor: LottoNumberGenerator를 도메인으로 변경
단순히 랜덤한 번호를 생성하는 것이 아니라 로또 규칙(6개, 1~45 범위 등)에 따라서 생성하므로 도메인에 가깝다고 생각함
- 패키지 util -> domain 이동
- static 메서드 -> 일반 메서드
